refactor: 웹 관련 패키지 이동
ALL 리팩토링에 따른 수정
